feat(sidebar): enhance perfomance
