+++ package.json
@@ -44,6 +44,7 @@
     ]
   },
   "devDependencies": {
+    "@types/react-window": "^1.8.8",
     "@types/styled-components": "^5.1.34"
   }
 }
