+++ src/components/common/SymbolSelect/styles.ts
@@ -1,0 +1,10 @@
+import { styled } from "styled-components";
+
+export const ListItem = styled.div`
+  padding: 14px 4px;
+
+  &:last-child {
+    margin-bottom: 0;
+    border-bottom: none;
+  }
+`;
